Move eventbus to be an internal package.
From types/eventbus to internal/evenbus.
No usage changes were required.
